Fix trigger of initial read for newly implemented nodes
The config node checks for intial reads right after the connection to the knx
bus was established. Nodes which register after the intital reads were triggered
will not be able to read the initial value.

Now the config node stores the current status and when a node is registered after
the status is `connected` it triggers the read request for this particular node.